requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pytesseract==0.3.10
opencv-python==4.7.0.72
Pillow==9.5.0
numpy==1.24.3
requests==2.31.0
google-cloud-storage==2.10.0
